# ProManage_API_Specification.yaml
openapi: 3.0.3
info:
  title: ProManage API
  description: |
    ProManage 项目管理系统 RESTful API 规范

    ## 功能概述
    - 统一知识库管理
    - 智能变更管理
    - 任务和测试用例管理
    - 实时通知和协作
    - 基于角色的访问控制

    ## 认证说明
    API 使用 JWT Bearer Token 进行认证。获取 token 后，在请求头中添加：
    ```
    Authorization: Bearer {your_jwt_token}
    ```

    ## 错误处理
    API 使用标准 HTTP 状态码，错误响应格式统一为：
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "错误描述",
        "details": {}
      }
    }
    ```

    ## 分页
    列表接口支持分页，参数包括：
    - `page`: 页码（从1开始）
    - `size`: 每页数量（默认20，最大100）
    - `sort`: 排序字段
    - `order`: 排序方向（asc/desc）

  version: 1.0.0
  contact:
    name: ProManage Development Team
    email: dev@promanage.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.promanage.com/v1
    description: 生产环境
  - url: https://staging-api.promanage.com/v1
    description: 测试环境
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
security:
  - BearerAuth: []
...
paths:
...
  /projects/{projectId}/tasks:
    get:
      tags: [Tasks]
      summary: 获取任务列表
      description: 获取项目的任务列表
      operationId: getTasks
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          description: 任务状态
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: priority
          in: query
          description: 任务优先级
          schema:
            $ref: '#/components/schemas/TaskPriority'
        - name: assignee_id
          in: query
          description: 指派人ID
          schema:
            type: integer
            format: int64
        - name: reporter_id
          in: query
          description: 报告人ID
          schema:
            type: integer
            format: int64
        - name: parent_task_id
          in: query
          description: 父任务ID
          schema:
            type: integer
            format: int64
        - name: labels
          in: query
          description: 标签筛选（逗号分隔）
          schema:
            type: string
        - name: due_date_from
          in: query
          description: 截止日期范围开始
          schema:
            type: string
            format: date
        - name: due_date_to
          in: query
          description: 截止日期范围结束
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'

    post:
      tags: [Tasks]
      summary: 创建任务
      description: 在项目中创建新任务
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: 任务创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: 获取任务详情
      description: 获取任务的详细信息
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'

    put:
      tags: [Tasks]
      summary: 更新任务
      description: 更新任务信息
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: 任务更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'

    delete:
      tags: [Tasks]
      summary: 删除任务
      description: 软删除任务
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '204':
          description: 任务删除成功
  /tasks/{taskId}/comments:
    get:
      tags: [Tasks]
      summary: 获取任务评论
      description: 获取任务的评论列表
      operationId: getTaskComments
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 评论列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TaskComment'

    post:
      tags: [Tasks]
      summary: 添加任务评论
      description: 为任务添加评论
      operationId: createTaskComment
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: 评论内容
                parent_comment_id:
                  type: integer
                  format: int64
                  description: 父评论ID（回复评论）
                is_internal:
                  type: boolean
                  description: 是否为内部评论
                  default: false
      responses:
        '201':
          description: 评论添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskComment'

  # ================================
  # 测试用例管理 API
  # ================================
...
  /analytics/tasks:
    get:
      tags: [Analytics]
      summary: 任务统计分析
      description: 获取任务相关的统计分析数据
      operationId: getTaskAnalytics
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
            format: int64
        - name: assignee_id
          in: query
          description: 指派人ID筛选
          schema:
            type: integer
            format: int64
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: weekly
      responses:
        '200':
          description: 任务统计数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAnalytics'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码（从1开始）
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 每页数量
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      description: 项目ID
      schema:
        type: integer
        format: int64
...
    TaskIdParam:
      name: taskId
      in: path
      required: true
      description: 任务ID
      schema:
        type: integer
        format: int64
...
  schemas:
...
    PaginatedResponse:
      type: object
      required: [data, pagination]
      properties:
        pagination:
          type: object
          required: [page, size, total, total_pages]
          properties:
            page:
              type: integer
              description: 当前页码
            size:
              type: integer
              description: 每页数量
            total:
              type: integer
              description: 总记录数
            total_pages:
              type: integer
              description: 总页数
            has_next:
              type: boolean
              description: 是否有下一页
            has_prev:
              type: boolean
              description: 是否有上一页

    # ================================
    # 枚举类型定义
    # ================================
...
    TaskStatus:
      type: string
      enum: [TODO, IN_PROGRESS, UNDER_REVIEW, COMPLETED, CANCELLED, BLOCKED]
      description: 任务状态
    TaskPriority:
      type: string
      enum: [LOW, NORMAL, HIGH, URGENT]
      description: 任务优先级
...
    Task:
      type: object
      required: [id, title, status, priority, reporter, project_id]
      properties:
        id:
          type: integer
          format: int64
          description: 任务ID
        title:
          type: string
          description: 任务标题
        description:
          type: string
          description: 任务描述
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        assignee:
          $ref: '#/components/schemas/User'
        reporter:
          $ref: '#/components/schemas/User'
        project_id:
          type: integer
          format: int64
          description: 项目ID
        parent_task_id:
          type: integer
          format: int64
          description: 父任务ID
        estimated_hours:
          type: number
          format: decimal
          description: 预估工时
        actual_hours:
          type: number
          format: decimal
          description: 实际工时
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
          description: 完成进度百分比
        start_date:
          type: string
          format: date
          description: 开始日期
        due_date:
          type: string
          format: date
          description: 截止日期
        tags:
          type: array
          items:
            type: string
          description: 标签
        labels:
          type: array
          items:
            type: string
          description: 标签
        created_at:
          type: string
          format: date-time
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          description: 更新时间
        completed_at:
          type: string
          format: date-time
          description: 完成时间
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            subtasks:
              type: array
              items:
                $ref: '#/components/schemas/Task'
              description: 子任务列表
            dependencies:
              type: array
              items:
                $ref: '#/components/schemas/Task'
              description: 依赖任务
            comments_count:
              type: integer
              description: 评论数量
            attachments_count:
              type: integer
              description: 附件数量
    TaskComment:
      type: object
      required: [id, content, author, created_at]
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
          description: 评论内容
        author:
          $ref: '#/components/schemas/User'
        parent_comment_id:
          type: integer
          format: int64
          description: 父评论ID
        is_internal:
          type: boolean
          description: 是否为内部评论
        created_at:
          type: string
          format: date-time
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          description: 更新时间
...
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 500
          description: 任务标题
        description:
          type: string
          description: 任务描述
        priority:
          $ref: '#/components/schemas/TaskPriority'
        assignee_id:
          type: integer
          format: int64
          description: 指派人ID
        parent_task_id:
          type: integer
          format: int64
          description: 父任务ID
        estimated_hours:
          type: number
          format: decimal
          description: 预估工时
        start_date:
          type: string
          format: date
          description: 开始日期
        due_date:
          type: string
          format: date
          description: 截止日期
        tags:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            type: string
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        assignee_id:
          type: integer
          format: int64
        estimated_hours:
          type: number
          format: decimal
        actual_hours:
          type: number
          format: decimal
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
        start_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            type: string
...
    TaskAnalytics:
      type: object
      properties:
        completion_trend:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              completed_count:
                type: integer
              total_count:
                type: integer
        priority_distribution:
          type: object
          additionalProperties:
            type: integer
        assignee_workload:
          type: array
          items:
            type: object
            properties:
              assignee:
                $ref: '#/components/schemas/User'
              total_tasks:
                type: integer
              completed_tasks:
                type: integer
              in_progress_tasks:
                type: integer
              average_completion_time:
                type: number
                description: 平均完成时间（天）
        overdue_tasks:
          type: object
          properties:
            count:
              type: integer
            percentage:
              type: number
              format: percentage
...